refactored out MemStream
